migrate handleEvent in the page instead of list components of user management & fix keys and lowercase letters in the button of the dialog
