+++ packages/playground/src/components/user-management/users-list.tsx
@@ -1,4 +1,4 @@
-import React, { useReducer } from 'react'
+import React from 'react'
 import {
   Icon,
   Table,
@@ -22,12 +22,7 @@ import {
   DropdownMenuTrigger
 } from '@harnessio/canary'
 import { getInitials } from '../../utils/utils'
-import { FormUserEditDialog } from './form-user-edit-dialog'
-import { FormRemoveAdminDialog } from './form-admin-remove-dialog'
-import { FormDeleteUserDialog } from './form-user-delete-dialog'
-import { FormResetPasswordDialog } from './form-user-reset-password'
 import { timeAgo } from '../../utils/utils'
-import { dialogStateReducer, initialDialogState } from './user-reducers/dialog-state-reducers'
 
 interface UsersProps {
   admin: boolean
@@ -42,72 +37,14 @@ interface UsersProps {
 
 interface PageProps {
   users: UsersProps[]
+  onDelete: (user: UsersProps) => void
+  onEdit: (user: UsersProps) => void
+  onRemoveAdmin: (user: UsersProps) => void
+  onResetPassword: (user: UsersProps) => void
 }
 
 // fix the edit form dialog and mock data and coressponding props
-export const UsersList = ({ users }: PageProps) => {
-  const [dialogState, dispatch] = useReducer(dialogStateReducer, initialDialogState)
-
-  const openDialog = (dialogType: 'delete' | 'edit' | 'removeAdmin' | 'resetPassword', user: UsersProps) => {
-    dispatch({ type: 'OPEN_DIALOG', dialogType, user })
-  }
-
-  const closeDialog = (dialogType: 'delete' | 'edit' | 'removeAdmin' | 'resetPassword') => {
-    dispatch({ type: 'CLOSE_DIALOG', dialogType })
-  }
-
-  // Delete user handler
-  const handleDelete = () => {
-    dispatch({ type: 'START_DELETING' })
-
-    // Simulate an API call
-    setTimeout(() => {
-      dispatch({ type: 'DELETE_SUCCESS' })
-      setTimeout(() => {
-        closeDialog('delete')
-        dispatch({ type: 'RESET_DELETE' })
-      }, 2000)
-    }, 2000)
-  }
-
-  //form edit submit
-  const handleFormSave = () => {
-    dispatch({ type: 'START_SUBMITTING' })
-
-    setTimeout(() => {
-      dispatch({ type: 'SUBMIT_SUCCESS' })
-      setTimeout(() => {
-        closeDialog('edit')
-        dispatch({ type: 'RESET_SUBMIT' })
-      }, 2000)
-    }, 2000)
-  }
-
-  // Delete project handler
-  const handleRemove = () => {
-    dispatch({ type: 'START_REMOVING' })
-
-    setTimeout(() => {
-      dispatch({ type: 'REMOVE_SUCCESS' })
-      setTimeout(() => {
-        closeDialog('removeAdmin')
-        dispatch({ type: 'RESET_REMOVE' })
-      }, 2000)
-    }, 2000)
-  }
-
-  // Reset password handler
-  const handleReset = () => {
-    dispatch({ type: 'START_RESETTING' })
-    setTimeout(() => {
-      dispatch({ type: 'RESET_PASSWORD_SUCCESS' })
-      setTimeout(() => {
-        closeDialog('resetPassword')
-        dispatch({ type: 'RESET_PASSWORD_RESET' })
-      }, 2000)
-    }, 2000)
-  }
-
+export const UsersList = ({ users, onDelete, onEdit, onRemoveAdmin, onResetPassword }: PageProps) => {
   const moreActionsTooltip = ({ user }: { user: UsersProps }) => {
     return (
       <DropdownMenu>
@@ -121,7 +58,7 @@ export const UsersList = ({ users }: PageProps) => {
             <DropdownMenuItem
               className="cursor-pointer"
               onSelect={() => {
-                openDialog('removeAdmin', user)
+                onRemoveAdmin(user)
               }}>
               <DropdownMenuShortcut className="ml-0">
                 <Icon name="trash" className="mr-2" />
@@ -131,7 +68,7 @@ export const UsersList = ({ users }: PageProps) => {
             <DropdownMenuItem
               className="cursor-pointer"
               onSelect={() => {
-                openDialog('resetPassword', user)
+                onResetPassword(user)
               }}>
               <DropdownMenuShortcut className="ml-0">
                 <Icon name="cog-6" className="mr-2" />
@@ -141,7 +78,7 @@ export const UsersList = ({ users }: PageProps) => {
             <DropdownMenuItem
               className="cursor-pointer"
               onSelect={() => {
-                openDialog('edit', user)
+                onEdit(user)
               }}>
               <DropdownMenuShortcut className="ml-0">
                 <Icon name="edit-pen" className="mr-2" />
@@ -152,7 +89,7 @@ export const UsersList = ({ users }: PageProps) => {
             <DropdownMenuItem
               className="cursor-pointer text-red-400 hover:text-red-400 focus:text-red-400"
               onSelect={() => {
-                openDialog('delete', user)
+                onDelete(user)
               }}>
               <DropdownMenuShortcut className="ml-0">
                 <Icon name="trash" className="mr-2 text-red-400" />
@@ -166,132 +103,80 @@ export const UsersList = ({ users }: PageProps) => {
   }
 
   return (
-    <>
-      <Table variant="asStackedList">
-        <TableHeader>
-          <TableRow>
-            <TableHead className="text-primary">Name</TableHead>
-            <TableHead className="text-primary">Email</TableHead>
-            <TableHead className="text-primary">Display Name</TableHead>
-            {users[0]?.created && <TableHead className="text-right text-primary">Date added</TableHead>}
-            <TableHead>
-              <></>
-            </TableHead>
-          </TableRow>
-        </TableHeader>
-        <TableBody>
-          {users &&
-            users.map(user => {
-              return (
-                <TableRow key={user.uid}>
-                  {/* NAME */}
-                  <TableCell className="content-center my-6">
-                    <div className="flex items-center gap-4">
-                      <Avatar size="10">
-                        {user.avatarUrl && <AvatarImage src={user.avatarUrl} />}
-                        <AvatarFallback className="text-xs p-1 text-center">{getInitials(user.uid, 2)}</AvatarFallback>
-                      </Avatar>
-                      <Text size={2} weight="medium" wrap="nowrap" truncate className="text-primary">
-                        {user.display_name}
-                        {user.admin && (
-                          <Badge
-                            variant="outline"
-                            size="xs"
-                            className="rounded-full font-normal text-xs p-2 h-5 text-tertiary-background text-center m-auto bg-tertiary-background/10 ml-2">
-                            Admin
-                          </Badge>
-                        )}
-                      </Text>
-                    </div>
-                  </TableCell>
-                  {/* EMAIL */}
-                  <TableCell className="content-center my-6">
-                    <div className="flex gap-1.5">
-                      <Text wrap="nowrap" size={1} truncate className="text-tertiary-background">
-                        {user.email}
-                      </Text>
-                    </div>
-                  </TableCell>
+    <Table variant="asStackedList">
+      <TableHeader>
+        <TableRow>
+          <TableHead className="text-primary">Name</TableHead>
+          <TableHead className="text-primary">Email</TableHead>
+          <TableHead className="text-primary">Display Name</TableHead>
+          <TableHead className="text-right text-primary">Date added</TableHead>
+          <TableHead>
+            <></>
+          </TableHead>
+        </TableRow>
+      </TableHeader>
+      <TableBody>
+        {users &&
+          users.map(user => {
+            return (
+              <TableRow key={user.uid}>
+                {/* NAME */}
+                <TableCell className="content-center my-6">
+                  <div className="flex items-center gap-4">
+                    <Avatar size="10">
+                      {user.avatarUrl && <AvatarImage src={user.avatarUrl} />}
+                      <AvatarFallback className="text-xs p-1 text-center">{getInitials(user.uid, 2)}</AvatarFallback>
+                    </Avatar>
+                    <Text size={2} weight="medium" wrap="nowrap" truncate className="text-primary">
+                      {user.display_name}
+                      {user.admin && (
+                        <Badge
+                          variant="outline"
+                          size="xs"
+                          className="rounded-full font-normal text-xs p-2 h-5 text-tertiary-background text-center m-auto bg-tertiary-background/10 ml-2">
+                          Admin
+                        </Badge>
+                      )}
+                    </Text>
+                  </div>
+                </TableCell>
+                {/* EMAIL */}
+                <TableCell className="content-center my-6">
+                  <div className="flex gap-1.5">
+                    <Text wrap="nowrap" size={1} truncate className="text-tertiary-background">
+                      {user.email}
+                    </Text>
+                  </div>
+                </TableCell>
 
-                  {/* displayName */}
-                  <TableCell className="content-center my-6">
-                    <div className="flex gap-1.5">
-                      <Text wrap="nowrap" size={1} truncate className="text-tertiary-background">
-                        {user.display_name}
-                      </Text>
-                    </div>
-                  </TableCell>
+                {/* displayName */}
+                <TableCell className="content-center my-6">
+                  <div className="flex gap-1.5">
+                    <Text wrap="nowrap" size={1} truncate className="text-tertiary-background">
+                      {user.display_name}
+                    </Text>
+                  </div>
+                </TableCell>
 
-                  {/* TimeStamp */}
-                  <TableCell className="content-center my-6">
-                    <div className="flex gap-1.5 items-center justify-end">
-                      <Text wrap="nowrap" size={1} truncate className="text-tertiary-background">
-                        {timeAgo(user.created)}
-                      </Text>
-                    </div>
-                  </TableCell>
+                {/* TimeStamp */}
+                <TableCell className="content-center my-6">
+                  <div className="flex gap-1.5 items-center justify-end">
+                    <Text wrap="nowrap" size={1} truncate className="text-tertiary-background">
+                      {timeAgo(user.created)}
+                    </Text>
+                  </div>
+                </TableCell>
 
-                  <TableCell className="content-center my-6">
-                    <div className="flex gap-1.5 items-center justify-end">
-                      {/* <Icon name="vertical-ellipsis" size={14} className="text-tertiary-background" /> */}
-                      {moreActionsTooltip({ user })}
-                    </div>
-                  </TableCell>
-                </TableRow>
-              )
-            })}
-        </TableBody>
-      </Table>
-      {/* Delete Dialog */}
-      {dialogState.isDialogDeleteOpen && (
-        <FormDeleteUserDialog
-          isDeleting={dialogState.isDeleting}
-          deleteSuccess={dialogState.deleteSuccess}
-          onDelete={handleDelete}
-          user={dialogState.selectedUser!}
-          onClose={() => {
-            closeDialog('delete')
-            dispatch({ type: 'RESET_DELETE' })
-          }}
-        />
-      )}
-      {/* Edit Dialog */}
-      {dialogState.isDialogEditOpen && (
-        <FormUserEditDialog
-          isSubmitting={dialogState.isSubmitting}
-          submitted={dialogState.submitted}
-          user={dialogState.selectedUser!}
-          onSave={handleFormSave}
-          onClose={() => {
-            closeDialog('edit')
-            dispatch({ type: 'RESET_SUBMIT' })
-          }}
-        />
-      )}
-      {dialogState.isDialogRemoveAdminOpen && (
-        <FormRemoveAdminDialog
-          isRemoving={dialogState.isRemoving}
-          removeSuccess={dialogState.removeSuccess}
-          user={dialogState.selectedUser!}
-          onRemove={handleRemove}
-          onClose={() => {
-            closeDialog('removeAdmin')
-            dispatch({ type: 'RESET_REMOVE' })
-          }}
-        />
-      )}
-      {dialogState.isDialogResetPasswordOpen && (
-        <FormResetPasswordDialog
-          isResetting={dialogState.isResetting}
-          resetSuccess={dialogState.resetSuccess}
-          user={dialogState.selectedUser!}
-          onReset={handleReset}
-          onClose={() => {
-            closeDialog('resetPassword')
-            dispatch({ type: 'RESET_PASSWORD_RESET' })
-          }}
-        />
-      )}
-    </>
+                <TableCell className="content-center my-6">
+                  <div className="flex gap-1.5 items-center justify-end">
+                    {/* <Icon name="vertical-ellipsis" size={14} className="text-tertiary-background" /> */}
+                    {moreActionsTooltip({ user })}
+                  </div>
+                </TableCell>
+              </TableRow>
+            )
+          })}
+      </TableBody>
+    </Table>
   )
 }
